// src/types/student-profile.ts
export interface StudentProfile {
    id: string
    user_id: string
    fullName: string
    email: string
    engineeringBranch: string
    collegeYear: string
    collegeNameAndLocation: string
    avatarUrl: string
    attendanceRate: number
    averageTestScore: number
    codingSkillScore: number
    communicationScore: number
    extracurricularScore: number
    leadershipScore: number
    internshipExperience: number
    activeBacklogs: string
    skills: string
    dream_companies: string
  }
  
  